% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morris-preprocessing.R
\name{as.navr}
\alias{as.navr}
\title{Converts the morris object to a navr object for easy plotting of paths and heatmaps}
\usage{
as.navr(obj_morris)
}
\arguments{
\item{obj_morris}{}
}
\description{
Converts the morris object to a navr object for easy plotting of paths and heatmaps
}
