% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morris-preprocessing.R
\name{combine_data_from_session}
\alias{combine_data_from_session}
\title{Combines data from multiple trials into a single navr object}
\usage{
combine_data_from_session(session)
}
\arguments{
\item{session}{session object as loaded by load_session}
}
\value{
single navr object with all position data together. 
Adds an iTrial column into the data for each separate trial. Strips away the settings
}
\description{
Combines data from multiple trials into a single navr object
}
