% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morris-loading.R
\name{load_session}
\alias{load_session}
\title{Loads all data from a single folder into a session object}
\usage{
load_session(folder)
}
\arguments{
\item{folder}{folder with multiple *.dat files}
}
\value{
session object list
}
\description{
Loads all data from a single folder into a session object
}
