% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morris-loading.R
\name{load_trial}
\alias{load_trial}
\title{Loads a single *.dat file into a single navr object}
\usage{
load_trial(filepath)
}
\arguments{
\item{filepath}{path to a valid *.dat file to load}
}
\value{
navr object
}
\description{
Loads a single *.dat file into a single navr object
}
