% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morris-visualising.R
\name{plot_heatmap.session}
\alias{plot_heatmap.session}
\title{Plots combined heatmap from multiple trials}
\usage{
plot_heatmap.session(session)
}
\arguments{
\item{session}{session object as loaded by load_session}
}
\description{
Plots combined heatmap from multiple trials
}
