% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morris-visualising.R
\name{plot_heatmap.trial}
\alias{plot_heatmap.trial}
\title{Plots a single trial heatmap}
\usage{
plot_heatmap.trial(obj)
}
\arguments{
\item{obj}{navr object as loaded by load_trial or from load_session}
}
\value{
ggplot heatmap
}
\description{
Plots a single trial heatmap
}
