% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morris-visualising.R
\name{plot_trial}
\alias{plot_trial}
\title{Plots a single trial with animal's path}
\usage{
plot_trial(obj)
}
\arguments{
\item{obj}{navr object}
}
\value{
ggplot object
}
\description{
Plots a single trial with animal's path
}
