% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morris-loading.R
\name{read_data}
\alias{read_data}
\title{Reads data into a unprocessed data.frame}
\usage{
read_data(filepath)
}
\arguments{
\item{filepath}{path to a valid *.dat file}
}
\value{
data.frame
}
\description{
unlike the load_trial, the loaded data is without the settings portion and not converted to the navr object
}
